Consolidate Decorator types and helpers
