+++ decorator/src/index.ts
@@ -1,12 +1,8 @@
 export type { IComms } from "@readium/navigator-html-injectables";
 export { Decorator } from "@readium/navigator-html-injectables";
 export type {
-    Decoration,
     DecoratorRequest,
     DecorationActivatedEvent as DecorationActivatedWireEvent,
-    DecorationStyle,
-    BuiltinDecorationStyle,
-    HTMLDecorationTemplate,
 } from "@readium/navigator-html-injectables";
 export {
     DecorationStyleType,
@@ -14,6 +10,7 @@ export {
     DecorationWidth,
 } from "@readium/navigator-html-injectables";
 
+export * from "./styles.ts";
 export * from "./comms/direct.ts";
 export * from "./Decoration.ts";
 export * from "./controller/DecorationController.ts";
